Use prev eye's base matrix
Experimental change that only fetches view matrix from the game on the first eye. For the second eye it will re-use that matrix, so camera ingame camera moves will not use two different offsets for each eye.

This experiment is not garanteed to sync with the eyes, and will need a way to account for that.